pyproject.toml
[project]
name = "helicorder-plotter"
version = "2.0.0"
description = "Seismic helicorder plot generator with interactive web output"
license = {text = "MIT"}
requires-python = ">=3.9"
dependencies = [
    "obspy>=1.4.0",
    "matplotlib>=3.5",
    "Pillow>=9.0",
    "aiohttp>=3.8",
    "pydantic>=2.0",
    "pyyaml>=6.0",
    "click>=8.0",
    "Jinja2>=3.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21",
    "aioresponses>=0.7",
]

[project.scripts]
helicorder = "helicorder.cli:main"

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["helicorder*"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
